requirements.txt
--extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/

# START section of dependencies that don't install on Darwin/MacOS
bitsandbytes==0.45.2
triton>=3.0.0
mamba-ssm==1.2.0.post1
flash-attn==2.7.4.post1
xformers>=0.0.23.post1
autoawq==0.2.7.post3
liger-kernel==0.5.3
# END section

packaging==23.2

peft==0.14.0
transformers==4.49.0
tokenizers>=0.21.0
accelerate==1.3.0
datasets==3.2.0
deepspeed==0.16.1
trl==0.15.1

optimum==1.16.2
hf_transfer
sentencepiece
gradio==3.50.2

modal==0.70.5
pydantic==2.10.6
addict
fire
PyYAML>=6.0
requests
wandb
einops
colorama
numba
numpy<2.0.1
# qlora things
evaluate==0.4.1
scipy
scikit-learn<1.4.3
nvidia-ml-py==12.560.30
art
tensorboard
python-dotenv==1.0.1

# remote filesystems
s3fs>=2024.5.0
gcsfs>=2024.5.0
# adlfs

zstandard<0.22.1
fastcore

# lm eval harness
lm_eval==0.4.7
langdetect==1.0.9
immutabledict==4.2.0
antlr4-python3-runtime==4.13.2

torchao==0.7.0
schedulefree==1.3.0

axolotl-contribs-lgpl==0.0.6
axolotl-contribs-mit==0.0.3
